Add files via upload
This C++ program demonstrates the execution of constructors and destructors in object-oriented programming. The Test class contains a user-defined constructor and destructor, which print messages when an object is created or destroyed. In the main function, an instance t of Test is created, triggering the constructor. When t goes out of scope at the end of main, the destructor is called. This showcases the automatic invocation of constructors and destructors.